Refactor: Radarr (#7088)
* Refactor

* Update radarr.sh

+++ ct/radarr.sh
@@ -29,19 +29,16 @@ function update_script() {
     exit
   fi
 
-  msg_info "Updating $APP LXC"
-  temp_file="$(mktemp)"
-  rm -rf /opt/Radarr
-  RELEASE=$(curl -fsSL https://api.github.com/repos/Radarr/Radarr/releases/latest | grep "tag_name" | awk '{print substr($2, 3, length($2)-4) }')
-  curl -fsSL "https://github.com/Radarr/Radarr/releases/download/v${RELEASE}/Radarr.master.${RELEASE}.linux-core-x64.tar.gz" -o "$temp_file"
-  $STD tar -xvzf "$temp_file"
-  mv Radarr /opt
-  chmod 775 /opt/Radarr
-  msg_ok "Updated $APP LXC"
+  RELEASE=$(curl -fsSL https://api.github.com/repos/Radarr/Radarr/releases/latest | jq -r '.tag_name' | sed 's/^v//')
+  if [[ ! -f ~/.radarr ]] || [[ "$RELEASE" != "$(cat ~/.radarr 2>/dev/null)" ]]; then
+    rm -rf /opt/Radarr
+    fetch_and_deploy_gh_release "Radarr" "Radarr/Radarr" "prebuild" "latest" "/opt/Radarr" "Radarr.master*linux-core-x64.tar.gz"
+    chmod 775 /opt/Radarr
+    msg_ok "Updated successfully"
+  else
+    msg_ok "No update required. $APP is already at v${RELEASE}"
+  fi
 
-  msg_info "Cleaning up"
-  rm -rf "$temp_file"
-  msg_ok "Cleaned up"
   exit
 }
